requirements.txt
pyopenjtalk-mod
tqdm